test(import-core): split import-engine.test.ts by engine phase
Splits the 2,195 LOC monolith into 6 focused files organized by engine
phase, plus a shared helper. Test count increases from 57 to 124.

Files:
  - import-engine-parsing.test.ts (222 LOC)
  - import-engine-ordering.test.ts (230 LOC)
  - import-engine-checkpoint.test.ts (523 LOC, 410 effective)
  - import-engine-persister.test.ts (562 LOC, 461 effective)
  - import-engine-persister-batch.test.ts (397 LOC)
  - import-engine-error-classification.test.ts (395 LOC)
  - helpers/engine-fixtures.ts (148 LOC)

Bean: ps-ga25
